requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
gunicorn==19.3.0
requests==2.7.0
python-google-places

